requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
python-dotenv
boto3>=1.42.55